Add files via upload
I created a project for the Hotel Management Network that aimed to improve efficiency and security. Key settings included OSPF for effective routing, inter-VLAN routing utilising the Router-on-a-Stick technique, and various VLANs for traffic segmentation and the hotel's multiple departments. In order to provide secure remote management, I implemented SSH access that was limited to the admin PC. In order to maximise address utilisation, exact IP subnetting was built inside a designated IP range, and port security was set up to prevent unauthorised host logins.